change route structure for a nested structure to reflect parent/child relationship

+++ app/Http/Controllers/ItemController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Box;
 use App\Models\Item;
 use App\Http\Requests\StoreItemRequest;
 use App\Http\Requests\UpdateItemRequest;
@@ -9,31 +10,39 @@ use Illuminate\Http\Response;
 
 class ItemController extends Controller
 {
-    public function index()
+    public function index(Box $box)
     {
         return response()->json([
-            'data' => Item::all()
+            'data' => $box->items
         ], Response::HTTP_OK);
     }
 
-    public function show(Item $item)
+    public function show(Box $box, Item $item)
     {
+        if ($item->box_id !== $box->id) {
+            abort(404);
+        }
+
         return response()->json([
             'data' => $item
         ], Response::HTTP_OK);
     }
 
-    public function store(StoreItemRequest $request)
+    public function store(StoreItemRequest $request, Box $box)
     {
-        $item = Item::create($request->validated());
+        $item = $box->items()->create($request->validated());
 
         return response()->json([
             'data' => $item
         ], Response::HTTP_CREATED);
     }
 
-    public function update(UpdateItemRequest $request, Item $item)
+    public function update(UpdateItemRequest $request, Box $box, Item $item)
     {
+        if ($item->box_id !== $box->id) {
+            abort(404);
+        }
+
         $item->update($request->validated());
 
         return response()->json([
@@ -41,8 +50,12 @@ class ItemController extends Controller
         ], Response::HTTP_OK);
     }
 
-    public function destroy(Item $item)
+    public function destroy(Box $box, Item $item)
     {
+        if ($item->box_id !== $box->id) {
+            abort(404);
+        }
+
         $item->delete();
 
         return response()->noContent(Response::HTTP_NO_CONTENT);
